+++ app/scripts/superdesk-authoring/tests/authoring_spec.js
@@ -7,12 +7,18 @@ describe('authoring', function() {
     var USER = 'user:1';
     var item = {guid: GUID};
 
+    beforeEach(module('superdesk.services.preferencesService'));
     beforeEach(module('superdesk.authoring'));
     beforeEach(module('superdesk.auth'));
     beforeEach(module('superdesk.mocks'));
 
     beforeEach(inject(function($window) {
         $window.onbeforeunload = angular.noop;
+    }));
+
+    beforeEach(inject(function(preferencesService, $q) {
+            spyOn(preferencesService, 'get').andReturn($q.when({'items':['urn:tag:superdesk-1']}));
+            spyOn(preferencesService, 'update').andReturn($q.when({}));
     }));
 
     beforeEach(inject(function($route) {
@@ -24,7 +30,7 @@ describe('authoring', function() {
         expect(session.identity._id).toBe(USER);
     }));
 
-    it('can open an item', inject(function(superdesk, api, lock, autosave, $injector, $q, $rootScope) {
+    it('can open an item', inject(function(superdesk, api, lock, autosave, $injector, $q, $rootScope, $httpBackend) {
         var _item,
             lockedItem = angular.extend({_locked: false}, item);