+++ app/scripts/superdesk/activity/superdesk-service.js
@@ -100,20 +100,23 @@ define([
             return this;
         };
 
-        this.$get = ['$q', '$rootScope', 'activityService', 'activityChooser', 'betaService', 'features',
-        function($q, $rootScope, activityService, activityChooser, betaService, features) {
+        this.$get = ['$q', '$route', '$rootScope', 'activityService', 'activityChooser', 'betaService', 'features',
+        function($q, $route, $rootScope, activityService, activityChooser, betaService, features) {
 
             /**
              * Render main menu depending on registered acitivites
              */
-            var beta = betaService.isBeta();
-            _.forEach(activities, function(activity, id) {
-                if (activity.beta === true && beta === false) {
-                    $routeProvider.when(activity.when, {redirectTo: '/workspace'});
-                    delete activities[id];
-                } else if (activity.when[0] === '/' && (activity.template || activity.templateUrl)) {
-                    $routeProvider.when(activity.when, activity);
-                }
+            betaService.isBeta().then(function(beta) {
+                _.forEach(activities, function(activity, id) {
+                    if (activity.beta === true && beta === false) {
+                        $routeProvider.when(activity.when, {redirectTo: '/workspace'});
+                        delete activities[id];
+                    } else if (activity.when[0] === '/' && (activity.template || activity.templateUrl)) {
+                        $routeProvider.when(activity.when, activity);
+                    }
+                });
+
+                $route.reload();
             });
 
             /**