+++ test-main.js
@@ -26,6 +26,7 @@ tests.push('superdesk/activity/activity');
 tests.push('superdesk/menu/notifications/notifications');
 tests.push('superdesk/services/translate');
 tests.push('superdesk/services/modalService');
+tests.push('superdesk/services/preferencesService');
 tests.push('superdesk/features/features');
 
 // apps