+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Superdesk Client Changelog
 
+## 0.3 (2014-02-24)
+
+- **users**: implement user form validation [ffdf478](https://github.com/superdesk/superdesk-client/commit/ffdf47834e5d7833098f03343b8d3e566df3e3d2)
+- **users**: check if username exists when creating a user [87b0f03](https://github.com/superdesk/superdesk-client/commit/87b0f0333eec5cb10107d5848e759622dc03294e)
+- **tests**: add protractor e2e test for login [d5bf1db](https://github.com/superdesk/superdesk-client/commit/d5bf1dbc2bfe4461090a6f575248b37532eb5813)
+
 ## 0.2 (2014-02-10)
 
 - feat(ingest): implement subject filtering